requirements.txt
astroid==2.8.6
attrs==21.2.0
black==21.9b0
cachetools==5.3.1
certifi==2021.10.8
charset-normalizer==3.1.0
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
google-api-core==2.11.0
google-auth==2.19.1
google-cloud-core==2.3.2
google-cloud-speech==2.20.0
google-cloud-storage==2.9.0
google-cloud-translate==3.11.1
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
gunicorn==20.1.0
idna==3.4
iniconfig==1.1.1
isort==5.12.0
itsdangerous==2.0.1
Jinja2==3.0.2
lazy-object-proxy==1.9.0
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.2
pathspec==0.9.0
pip==23.0.1
platformdirs==2.4.0
pluggy==1.0.0
proto-plus==1.22.2
protobuf==4.23.2
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydub==0.25.1
pylint==2.11.1
pyparsing==2.4.7
pytest==6.2.5
python-dotenv==1.0.0
pytube==15.0.0
regex==2021.10.23
requests==2.31.0
rsa==4.9
setuptools==67.8.0
six==1.16.0
srt==3.5.3
toml==0.10.2
tomli==1.2.2
typing-extensions==3.10.0.2
urllib3==1.26.16
Werkzeug==2.0.2
wheel==0.38.4
wrapt==1.13.3
